lint: move reply tray platform gates into shell files
